Add files via upload
The main goal for this project is to analize data from App Store and Google Play Market in order to help a company that builds free apps to make data-driven decision. The company has revenue for in-app adds, so the amount of revenue depends on how many users has our app.